requirements.txt
captum==0.6.0
cartopy==0.22.0
matplotlib==3.5.1
mayavi==4.8.1
mctorch==0.1.0
mctorch_lib==0.1.0
networkx==2.7.1
numpy==1.21.5
opencv_python==4.8.0.74
pandas==2.0.3
Pillow==9.0.1
pycocotools==2.0.7
pytorch_revgrad==0.2.0
scikit_learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
setuptools==61.2.0
tikzplotlib==0.10.1
torch==1.13.0
torchvision==0.15.2
tqdm==4.64.0
transformers==4.30.2
twokenize==1.0.0



